Small enhancement.
#69 #71 #99 #101 
Use the `is` operator instead of `==` to compare to `None`, because `==` calls `__eq__` which is slower and error-prone.

#66 #96
To check if a variable is equal to one of many values, combine the values into a tuple and check if the variable is contained in it instead of checking for equality against each of the values. This is faster, less verbose, and more readable.